refactor(hybrid): migrate hybrid functions to another module
